split out checks from ufo.py
